+++ svelte/src/lib/components/battery.svelte
