+++ svelte/src/lib/components/roomWidget.svelte
