+++ svelte/src/lib/components/sensor_icon.svelte
