+++ svelte/src/lib/components/update_icon.svelte
