+++ svelte/src/lib/components/weatherWidget.svelte
@@ -1,0 +1,336 @@
+<script lang="ts">
+    import { onMount } from "svelte";
+  
+    let data: { current: { temp: any; feels_like: number; humidity: number; weather: { icon: any; }[]; }; hourly: any; } | null;
+  
+  
+    let currentTime = ()=>
+    {
+          
+      let date = new Date();
+      let hours = date.getHours();
+      let minutes = date.getMinutes();
+      let seconds = date.getSeconds();
+      let day = date.getDate();
+      let month = date.getMonth() + 1;
+      let year = date.getFullYear();
+  
+      let returnTime:any={};
+      returnTime[hours]= hours;
+      returnTime[minutes]= minutes;
+      returnTime[seconds]= seconds;
+      returnTime[day]= day;
+      returnTime[month]= month;
+      returnTime[year]= year;
+      
+  
+  
+      return returnTime;
+    }
+  
+    let convertTimestamp = (timestamp: any, type: string) =>
+    {
+  
+      let date = new Date(timestamp * 1000);
+      let hours = date.getHours();
+      let minutes = date.getMinutes();
+      let seconds = date.getSeconds();
+      let day = date.getDate();
+      let month = date.getMonth() + 1;
+      let year = date.getFullYear();
+  
+      if (type == "date")
+      {
+          return `${day}.${month}.${year}`;
+      }
+      else if (type == "time"){
+          return `${hours}:${minutes}:${seconds}`;
+      }
+      else if (type == "day"){
+          return `${day}`;
+      }
+      else if (type == "month"){
+          return `${month}`;
+      }
+      else if (type == "year"){
+          return `${year}`;
+      }
+      else if (type == "hours"){
+          return `${hours}`;
+      }
+      else if (type == "minutes"){
+          return `${minutes}`;
+      }
+      else if (type == "seconds"){
+          return `${seconds}`;
+      }else return false;
+    }
+  
+    let checkTimestampHours = (timsestamp: number, hoursNext: any) =>
+    {
+      let date = new Date(timsestamp * 1000);
+      let hours = date.getHours();
+      let minutes = date.getMinutes();
+      
+  
+      let currentTimev = currentTime();
+          if(convertTimestamp(timsestamp, "hours") <=  currentTimev.hours+ hoursNext)
+          {
+              return true;
+          }
+          else
+          {
+              return false;
+          }
+  
+    }
+  
+  
+    // get forecast
+    function getForecast() {
+      fetch(
+      //   "https://weather.visualcrossing.com/VisualCrossingWebServices/rest/services/timeline/klecza%20dolna%2C%2034-124%20Poland%2C%20%20ul.spacerowa%20324/today?unitGroup=metric&elements=datetime%2Cname%2Caddress%2Ctempmax%2Ctempmin%2Ctemp%2Cfeelslike%2Chumidity%2Csnow%2Cwindspeed%2Cpressure%2Ccloudcover%2Cvisibility%2Csunrise%2Csunset%2Cicon&include=current%2Cdays%2Chours%2Calerts&key=5CVKAYPUXP5MPJ84S7E4L49QF&contentType=json"
+          // "https://api.openweathermap.org/data/2.5/weather?lat=49.880870&lon=19.563030&appid=f055d509de51700a688e61d5f8e3da76&units=metric"
+      // new API 
+      "https://api.openweathermap.org/data/3.0/onecall?lat=49.880870&lon=19.563030&appid=c9eb7dafaf1068ad6654df45dee8ae81&units=metric"
+      // (49.880870, 19.563030)
+          )
+        .then((response) => response.json())
+        .then((data_) => {
+          data = data_;
+        });
+    }
+  
+    onMount(async () => {
+      getForecast();
+    });
+  </script>
+  
+  {#if data != null}
+    <div
+      class="weather_widget flex flex-col items-center justify-center w-screen min-h-screen text-gray-700 p-10 "
+    >
+      <!-- Component Start -->
+      <div
+        class="w-full max-w-screen-sm bg-white p-10 rounded-xl ring-8 ring-white ring-opacity-40"
+      >
+        <div class="flex justify-between">
+          <div class="flex flex-col">
+            <span class="text-6xl font-bold lcd"
+              >{data.current.temp}<span class="text-3xl">°C</span></span
+            >
+            <span class="font-light mt-1 text-gray-500"
+              >Odczywalna : {data.current.feels_like.toFixed(1)}</span
+            >
+            <!-- <span class="font-semibold mt-1 text-gray-500">{data.name}</span> -->
+          </div>
+          <div class="flex flex-col">
+            <span class="text-6xl font-bold lcd"
+              >{data.current.humidity.toFixed(0)}<span class="text-3xl"
+                ><span class="upperTop">o</span> /o</span
+              ></span>
+           
+            <!-- <span class="font-semibold mt-1 text-gray-500">{data.name}</span> -->
+          </div>
+          <div class="h-24 w-24 ">
+              <!-- {data.current.weather[0].icon} -->
+            <img
+              src="http://openweathermap.org/img/wn/{data.current.weather[0].icon}@2x.png"
+              alt=""
+            />
+          </div>
+        </div>
+            <div class="flex justify-between mt-12">
+  
+                  {#each data.hourly as nexthour}
+                      <!-- {nexthour.dt} -->
+                      {convertTimestamp(nexthour.dt, "hours")}
+                      
+                      {#if checkTimestampHours(nexthour.dt, 6) == true}
+                          <div class="flex flex-col items-center">
+                              <span class="font-semibold text-lg lcd">{nexthour.temp.toFixed(0)}°C</span>
+                              <img alt="" src="http://openweathermap.org/img/wn/{nexthour.weather[0].icon}.png" />	
+                              <span class="font-semibold mt-1 text-sm lcd">{nexthour.dt}</span>
+                              <span class="text-xs font-semibold text-gray-400 lcd">PM</span>
+                          </div>
+                      {/if}
+                      
+                  {/each }
+  
+              
+          </div>
+      </div>
+  
+      <div
+        class="flex flex-col space-y-6 w-full max-w-screen-sm bg-white p-10 mt-10 rounded-xl ring-8 ring-white ring-opacity-40"
+      >
+        <div class="flex justify-between items-center">
+          <span class="font-semibold text-lg w-1/4">Fri, 22 Jan</span>
+          <div class="flex items-center justify-end w-1/4 pr-10">
+            <span class="font-semibold">12%</span>
+            <svg
+              class="w-6 h-6  ml-1"
+              viewBox="0 0 16 20"
+              version="1.1"
+              xmlns="http://www.w3.org/2000/svg"
+            >
+              <g transform="matrix(1,0,0,1,-4,-2)">
+                <path
+                  d="M17.66,8L12.71,3.06C12.32,2.67 11.69,2.67 11.3,3.06L6.34,8C4.78,9.56 4,11.64 4,13.64C4,15.64 4.78,17.75 6.34,19.31C7.9,20.87 9.95,21.66 12,21.66C14.05,21.66 16.1,20.87 17.66,19.31C19.22,17.75 20,15.64 20,13.64C20,11.64 19.22,9.56 17.66,8ZM6,14C6.01,12 6.62,10.73 7.76,9.6L12,5.27L16.24,9.65C17.38,10.77 17.99,12 18,14C18.016,17.296 14.96,19.809 12,19.74C9.069,19.672 5.982,17.655 6,14Z"
+                  style="fill-rule:nonzero;"
+                />
+              </g>
+            </svg>
+          </div>
+          <svg
+            class="h-8 w-8  w-1/4"
+            xmlns="http://www.w3.org/2000/svg"
+            height="24"
+            viewBox="0 0 24 24"
+            width="24"
+            ><path d="M0 0h24v24H0V0z" fill="none" /><path
+              d="M6.76 4.84l-1.8-1.79-1.41 1.41 1.79 1.79zM1 10.5h3v2H1zM11 .55h2V3.5h-2zm8.04 2.495l1.408 1.407-1.79 1.79-1.407-1.408zm-1.8 15.115l1.79 1.8 1.41-1.41-1.8-1.79zM20 10.5h3v2h-3zm-8-5c-3.31 0-6 2.69-6 6s2.69 6 6 6 6-2.69 6-6-2.69-6-6-6zm0 10c-2.21 0-4-1.79-4-4s1.79-4 4-4 4 1.79 4 4-1.79 4-4 4zm-1 4h2v2.95h-2zm-7.45-.96l1.41 1.41 1.79-1.8-1.41-1.41z"
+            /></svg
+          >
+          <span class="font-semibold text-lg w-1/4 text-right">18° / 32°</span>
+        </div>
+        <div class="flex justify-between items-center">
+          <span class="font-semibold text-lg w-1/4">Sat, 23 Jan</span>
+          <div class="flex items-center justify-end pr-10 w-1/4">
+            <span class="font-semibold">0%</span>
+            <svg
+              class="w-6 h-6  ml-1"
+              viewBox="0 0 16 20"
+              version="1.1"
+              xmlns="http://www.w3.org/2000/svg"
+            >
+              <g transform="matrix(1,0,0,1,-4,-2)">
+                <path
+                  d="M17.66,8L12.71,3.06C12.32,2.67 11.69,2.67 11.3,3.06L6.34,8C4.78,9.56 4,11.64 4,13.64C4,15.64 4.78,17.75 6.34,19.31C7.9,20.87 9.95,21.66 12,21.66C14.05,21.66 16.1,20.87 17.66,19.31C19.22,17.75 20,15.64 20,13.64C20,11.64 19.22,9.56 17.66,8ZM6,14C6.01,12 6.62,10.73 7.76,9.6L12,5.27L16.24,9.65C17.38,10.77 17.99,12 18,14C18.016,17.296 14.96,19.809 12,19.74C9.069,19.672 5.982,17.655 6,14Z"
+                  style="fill-rule:nonzero;"
+                />
+              </g>
+            </svg>
+          </div>
+          <svg
+            class="h-8 w-8  w-1/4"
+            xmlns="http://www.w3.org/2000/svg"
+            height="24"
+            viewBox="0 0 24 24"
+            width="24"
+            ><path d="M0 0h24v24H0V0z" fill="none" /><path
+              d="M6.76 4.84l-1.8-1.79-1.41 1.41 1.79 1.79zM1 10.5h3v2H1zM11 .55h2V3.5h-2zm8.04 2.495l1.408 1.407-1.79 1.79-1.407-1.408zm-1.8 15.115l1.79 1.8 1.41-1.41-1.8-1.79zM20 10.5h3v2h-3zm-8-5c-3.31 0-6 2.69-6 6s2.69 6 6 6 6-2.69 6-6-2.69-6-6-6zm0 10c-2.21 0-4-1.79-4-4s1.79-4 4-4 4 1.79 4 4-1.79 4-4 4zm-1 4h2v2.95h-2zm-7.45-.96l1.41 1.41 1.79-1.8-1.41-1.41z"
+            /></svg
+          >
+          <span class="font-semibold text-lg w-1/4 text-right">22° / 34°</span>
+        </div>
+        <div class="flex justify-between items-center">
+          <span class="font-semibold text-lg w-1/4">Sun, 24 Jan</span>
+          <div class="flex items-center justify-end pr-10 w-1/4">
+            <span class="font-semibold">20%</span>
+            <svg
+              class="w-6 h-6  ml-1"
+              viewBox="0 0 16 20"
+              version="1.1"
+              xmlns="http://www.w3.org/2000/svg"
+            >
+              <g transform="matrix(1,0,0,1,-4,-2)">
+                <path
+                  d="M17.66,8L12.71,3.06C12.32,2.67 11.69,2.67 11.3,3.06L6.34,8C4.78,9.56 4,11.64 4,13.64C4,15.64 4.78,17.75 6.34,19.31C7.9,20.87 9.95,21.66 12,21.66C14.05,21.66 16.1,20.87 17.66,19.31C19.22,17.75 20,15.64 20,13.64C20,11.64 19.22,9.56 17.66,8ZM6,14C6.01,12 6.62,10.73 7.76,9.6L12,5.27L16.24,9.65C17.38,10.77 17.99,12 18,14C18.016,17.296 14.96,19.809 12,19.74C9.069,19.672 5.982,17.655 6,14Z"
+                  style="fill-rule:nonzero;"
+                />
+              </g>
+            </svg>
+          </div>
+          <svg
+            class="h-8 w-8  w-1/4"
+            xmlns="http://www.w3.org/2000/svg"
+            height="24"
+            viewBox="0 0 24 24"
+            width="24"
+            ><path d="M0 0h24v24H0V0z" fill="none" /><path
+              d="M6.76 4.84l-1.8-1.79-1.41 1.41 1.79 1.79zM1 10.5h3v2H1zM11 .55h2V3.5h-2zm8.04 2.495l1.408 1.407-1.79 1.79-1.407-1.408zm-1.8 15.115l1.79 1.8 1.41-1.41-1.8-1.79zM20 10.5h3v2h-3zm-8-5c-3.31 0-6 2.69-6 6s2.69 6 6 6 6-2.69 6-6-2.69-6-6-6zm0 10c-2.21 0-4-1.79-4-4s1.79-4 4-4 4 1.79 4 4-1.79 4-4 4zm-1 4h2v2.95h-2zm-7.45-.96l1.41 1.41 1.79-1.8-1.41-1.41z"
+            /></svg
+          >
+          <span class="font-semibold text-lg w-1/4 text-right">21° / 32°</span>
+        </div>
+        <div class="flex justify-between items-center">
+          <span class="font-semibold text-lg w-1/4">Mon, 25 Jan</span>
+          <div class="flex items-center justify-end pr-10 w-1/4">
+            <span class="font-semibold">50%</span>
+            <svg
+              class="w-6 h-6  ml-1"
+              viewBox="0 0 16 20"
+              version="1.1"
+              xmlns="http://www.w3.org/2000/svg"
+            >
+              <g transform="matrix(1,0,0,1,-4,-2)">
+                <path
+                  d="M17.66,8L12.71,3.06C12.32,2.67 11.69,2.67 11.3,3.06L6.34,8C4.78,9.56 4,11.64 4,13.64C4,15.64 4.78,17.75 6.34,19.31C7.9,20.87 9.95,21.66 12,21.66C14.05,21.66 16.1,20.87 17.66,19.31C19.22,17.75 20,15.64 20,13.64C20,11.64 19.22,9.56 17.66,8ZM6,14C6.01,12 6.62,10.73 7.76,9.6L12,5.27L16.24,9.65C17.38,10.77 17.99,12 18,14C18.016,17.296 14.96,19.809 12,19.74C9.069,19.672 5.982,17.655 6,14Z"
+                  style="fill-rule:nonzero;"
+                />
+              </g>
+            </svg>
+          </div>
+          <svg
+            class="h-8 w-8  w-1/4"
+            xmlns="http://www.w3.org/2000/svg"
+            height="24"
+            viewBox="0 0 24 24"
+            width="24"
+            ><path d="M0 0h24v24H0V0z" fill="none" /><path
+              d="M12.01 6c2.61 0 4.89 1.86 5.4 4.43l.3 1.5 1.52.11c1.56.11 2.78 1.41 2.78 2.96 0 1.65-1.35 3-3 3h-13c-2.21 0-4-1.79-4-4 0-2.05 1.53-3.76 3.56-3.97l1.07-.11.5-.95C8.08 7.14 9.95 6 12.01 6m0-2C9.12 4 6.6 5.64 5.35 8.04 2.35 8.36.01 10.91.01 14c0 3.31 2.69 6 6 6h13c2.76 0 5-2.24 5-5 0-2.64-2.05-4.78-4.64-4.96C18.68 6.59 15.65 4 12.01 4z"
+            /></svg
+          >
+          <span class="font-semibold text-lg w-1/4 text-right">18° / 29°</span>
+        </div>
+        <div class="flex justify-between items-center">
+          <span class="font-semibold text-lg w-1/4">Tue, 26 Jan</span>
+          <div class="flex items-center justify-center w-1/4">
+            <span class="font-semibold">80%</span>
+            <svg
+              class="w-6 h-6  ml-1"
+              viewBox="0 0 16 20"
+              version="1.1"
+              xmlns="http://www.w3.org/2000/svg"
+            >
+              <g transform="matrix(1,0,0,1,-4,-2)">
+                <path
+                  d="M17.66,8L12.71,3.06C12.32,2.67 11.69,2.67 11.3,3.06L6.34,8C4.78,9.56 4,11.64 4,13.64C4,15.64 4.78,17.75 6.34,19.31C7.9,20.87 9.95,21.66 12,21.66C14.05,21.66 16.1,20.87 17.66,19.31C19.22,17.75 20,15.64 20,13.64C20,11.64 19.22,9.56 17.66,8ZM6,14C6.01,12 6.62,10.73 7.76,9.6L12,5.27L16.24,9.65C17.38,10.77 17.99,12 18,14C18.016,17.296 14.96,19.809 12,19.74C9.069,19.672 5.982,17.655 6,14Z"
+                  style="fill-rule:nonzero;"
+                />
+              </g>
+            </svg>
+          </div>
+          <svg
+            class="h-8 w-8  w-1/4"
+            xmlns="http://www.w3.org/2000/svg"
+            height="24"
+            viewBox="0 0 24 24"
+            width="24"
+            ><path d="M0 0h24v24H0V0z" fill="none" /><path
+              d="M12.01 6c2.61 0 4.89 1.86 5.4 4.43l.3 1.5 1.52.11c1.56.11 2.78 1.41 2.78 2.96 0 1.65-1.35 3-3 3h-13c-2.21 0-4-1.79-4-4 0-2.05 1.53-3.76 3.56-3.97l1.07-.11.5-.95C8.08 7.14 9.95 6 12.01 6m0-2C9.12 4 6.6 5.64 5.35 8.04 2.35 8.36.01 10.91.01 14c0 3.31 2.69 6 6 6h13c2.76 0 5-2.24 5-5 0-2.64-2.05-4.78-4.64-4.96C18.68 6.59 15.65 4 12.01 4z"
+            /></svg
+          >
+          <span class="font-semibold text-lg w-1/4 text-right">20° / 29°</span>
+        </div>
+      </div>
+      <!-- Component End  -->
+    </div>
+  {/if}
+  
+  <style>
+    @font-face {
+      font-family: "lcd";
+      src: url("././aclock.ttf") format("truetype"); /* Safari, Android, iOS */
+    }
+  
+    .weather_widget .lcd {
+      font-family: "lcd", Fallback, sans-serif !important;
+    }
+    .upperTop {
+      position: absolute;
+      margin-top: 8px;
+      margin-left: 7px;
+    }
+  </style>
